requirements.txt
flask
flask-cors
pymysql
python-dotenv
pydantic
pydantic-settings
google-cloud-firestore
cryptography
python-telegram-bot
passlib[bcrypt]
python-jose[cryptography]
